+++ src/history.hpp
@@ -33,7 +33,7 @@ public:
     // Change the score of a move which failed to cause a beta-cutoff at the given depth when some other move did cause one.
     void addNotCutoff(const Position& pos, const Move& move, int depth);
     // Get the history score for the given move.
-    int getScore(const Position& pos, const Move& move) const;
+    int16_t getScore(const Position& pos, const Move& move) const;
     // Clears the table.
     void clear();
     // From time to time we should age the table to make sure that scores for previous positions won't dominate the scores.
